add user view, add user stats, fix linux bugs
- add user view: includes all matches tracked, user skill, user map
stats, user faction stats
- add user stats: include per-gamemode skill, map stats and faction
stats
- allow image proxy when coloring an image to use a non-colored one
that already exists
- add gamemode to BAR match
- linux headless: fix assumptions about where files are compared
to windows
- headless: enable multiple instances to be running, uses
different ports and data folders
- add player ID to spectators
- update demofile parsing to include faction swaps from the
gadget (enabled legion and random)
- match view: add heat maps for unit deaths, commander
position and static unit creation
- match view: rotate factories (matters for air as it's a
rectangle)
- add legion icons

+++ gex/Models/Db/BarMatchPlayer.cs
@@ -48,12 +48,12 @@ namespace gex.Models.Db {
         /// <summary>
         ///     skill value at the time of this match
         /// </summary>
-        public decimal Skill { get; set; }
+        public double Skill { get; set; }
 
         /// <summary>
         ///     uncertainty in the skill at the time of the match
         /// </summary>
-        public decimal SkillUncertainty { get; set; }
+        public double SkillUncertainty { get; set; }
 
         /// <summary>
         ///     32 bit representation of the color, (r &lt;&lt; 16) | (g &lt;&lt; 8) | (b &lt;&lt; 0)
